+++ src/runtime_src/xdp/profile/database/events/vtf_event.h
@@ -111,7 +111,7 @@ namespace xdp {
 	                                    type == LOP_KERNEL_ENQUEUE ; }
 
     virtual uint64_t getDevice() { return 0 ; } // CHECK
-    XDP_EXPORT virtual void dump(std::ofstream& fout, int bucket) ;
+    XDP_EXPORT virtual void dump(std::ofstream& fout, uint64_t bucket) ;
   } ;
 
   // Used so the database can sort based on timestamp order
